Moved `app/KernelLoader.php` to `App\Component\Application`
