+++ src/Backend/Core/Engine/AjaxAction.php
@@ -4,7 +4,7 @@ namespace Backend\Core\Engine;
 
 use Backend\Core\Engine\Base\Config;
 use Backend\Core\Engine\Base\AjaxAction as BaseAjaxAction;
-use ForkCMS\App\KernelLoader;
+use App\Component\Application\KernelLoader;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\KernelInterface;
