+++ src/Backend/Core/Engine/Base/Action.php
@@ -4,7 +4,7 @@ namespace Backend\Core\Engine\Base;
 
 use Backend\Core\Engine\TwigTemplate;
 use App\Exception\RedirectException;
-use ForkCMS\App\KernelLoader;
+use App\Component\Application\KernelLoader;
 use Symfony\Component\Form\Form;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
