+++ src/Backend/Core/Engine/Base/AjaxAction.php
@@ -3,7 +3,7 @@
 namespace Backend\Core\Engine\Base;
 
 use App\Component\Model\BackendModel;
-use ForkCMS\App\KernelLoader;
+use App\Component\Application\KernelLoader;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
