+++ src/Backend/Core/Engine/Base/Widget.php
@@ -4,7 +4,7 @@ namespace Backend\Core\Engine\Base;
 
 use App\Component\Model\BackendModel;
 use App\Exception\RedirectException;
-use ForkCMS\App\KernelLoader;
+use App\Component\Application\KernelLoader;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
