+++ src/Backend/Core/Engine/Navigation.php
@@ -2,7 +2,7 @@
 
 namespace Backend\Core\Engine;
 
-use ForkCMS\App\KernelLoader;
+use App\Component\Application\KernelLoader;
 use Symfony\Component\HttpKernel\KernelInterface;
 use App\Component\Model\BackendModel;
 use Backend\Modules\Pages\Engine\Model as BackendPagesModel;
