+++ src/Backend/Core/Engine/Url.php
@@ -6,7 +6,7 @@ use Backend\Core\Config;
 use Backend\Core\Engine\Base\Config as BackendBaseConfig;
 use App\Component\Model\BackendModel;
 use App\Exception\RedirectException;
-use ForkCMS\App\KernelLoader;
+use App\Component\Application\KernelLoader;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
