+++ src/Component/Application/Init.php
@@ -3,7 +3,7 @@
 namespace App\Component\Application;
 
 use App\Exception\InvalidInitTypeException;
-use ForkCMS\App\KernelLoader;
+use App\Component\Application\KernelLoader;
 
 /**
  * This class will initiate the application
