+++ src/Component/Application/KernelLoader.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace ForkCMS\App;
+namespace App\Component\Application;
 
-use Symfony\Component\DependencyInjection\ContainerInterface;
-use Symfony\Component\HttpKernel\KernelInterface;
 use App\Component\Model\BackendModel;
 use App\Component\Model\FrontendModel;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+use Symfony\Component\HttpKernel\KernelInterface;
 
 /**
  * This class is used in several Fork applications to bubble down the AppKernel/Kernel object.
