+++ src/Frontend/Core/Engine/Base/Config.php
@@ -2,7 +2,7 @@
 
 namespace Frontend\Core\Engine\Base;
 
-use ForkCMS\App\KernelLoader;
+use App\Component\Application\KernelLoader;
 use Symfony\Component\HttpKernel\KernelInterface;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Finder\Finder;
